+++ lucadeal/lib/luca_deal/contract.rb
@@ -43,6 +43,16 @@ module LucaDeal
       end
     end
 
+    def describe(id)
+      contract = parse_current(self.class.find(id))
+      if contract['products']
+        contract['products'] = contract['products'].map do |product|
+          LucaDeal::Product.find(product['id'])
+        end
+      end
+      YAML.dump(readable(contract)).tap{ |d| puts d }
+    end
+
     def generate!(customer_id, mode = 'subscription')
       LucaDeal::Customer.find(customer_id) do |customer|
         current_customer = parse_current(customer)